Set up add-delete for postgres storage
At this point we have a copy of the flipper ActiveRecord schema, and can
add/remove features and have them persisted, though not actually do
anything with them (same as redis/in-memory).

Revised kaocha, deps.edn paths to --watch properly, and tests now rely
on `:extra-deps` instead of `:paths`, which drags in the right deps
(carmine, psql). /thx [Alys][1]

[1]: https://clojurians.slack.com/archives/CCY2V0U6A/p1675106292376889?thread_ts=1674847481.340129&cid=CCY2V0U6A